Grid Header
Changed the header to a display: grid with three columns using the fr value so they are all the same size. Therefore my title is centered and my nav bar is to the right. Also justified content in nav bar to the flex end so it appears to the right hand side.