planet buttons
Created 3 buttons for the user to press. Each button will open the article section underneath and the content of this article section will be filled using javascript. This will allow me control over the content in this section. I could add some css to the article in order to manipulate the appearence but at the moment it looks like I want it to in block