planet_buttons
created a forEach that loops through each button. made the function anonymous so it doesn't get called on opening the page or refreshing etc. created two variables to control whether the content should be updated or closed/opened. passing button_id as an argument into the function which is retrieved from the previous loop. if else conditional checks if content is open && current_id matches button id. If they do match then the content is closed as user has pressed the same button. if not the runs else function which uses .innerHTML to add the content to the webpage.