Update week_1_final_proj.html
added hidden class to contact section which can be added and removed in eventListener. Also placed the close icon within a div as it was spanning the whole length of the contact form. Now div spans the whole length and justifies the content the flex-end. Close button no longer spans entire width.