Update final_proj.js
added event listener on close button that adds hidden to class of contact section - therefore hiding the section or 'closing' it. Also added event listeners to the a_buttons on the page. needed to reference the event as default a event refreshes page which meant the buttons were not acting as anticipated. just used preventdefualt() built in function to stop the a_buttons defualt action. also removes the hidden class list therefore 'opening' the contact form